Delete user reputation if they already exist in the database
The code deletes a user's reputation from the database if they already exist, before creating a new one with the same telegramId. This ensures that each user only has one reputation

+++ src/bot/bot.service.ts
@@ -65,12 +65,15 @@ export class BotService implements OnModuleInit {
         const user = await this.prisma.reputations.findFirst({
             where: { telegramId },
         })
-
-        await this.prisma.reputations.delete({
-            where: {
-                id: user.id,
-            },
-        })
+        if (user) {
+            await this.prisma.reputations.delete({
+                where: {
+                    id: user.id,
+                },
+            })
+        } else {
+            return
+        }
     }
 
     async getReputation(telegramId: string): Promise<Reputations> {
